add support for numberMatched in Orders response (#123)
## What I'm changing

- RootRouter Callable `get_orders` now requires an additional value in
the result tuple that is a count of the total number of orders
that will be returned from pagination. When this returns `Some(int)`,
the value is used for the `numberMatched` field in
FeatureCollection returned from the /orders endpoint. If this feature is
not desired, providing a function that returns
  `Nothing` will exclude the `numberMatched` field in the response.
- Removed dependency on `pyrfc3339` library, since only one function
from it was used in tests and that function has
  been removed in newer versions of the library.
- pydantic >= 2.12 is now required.
- bump stapi-pydantic to 0.1.0, bump stapi-fastapi to 0.8.0

## How I did it

- Changed the signature on get_orders instead of adding another callable
like get_orders_count. This makes it a bit cleaner. The 3-tuple returned
by get_orders should be changed to a class if any more values are added.

## Checklist

- [X] Tests pass: `uv run pytest`
- [X] Checks pass: `uv run pre-commit run --all-files`
- [X] CHANGELOG is updated (if necessary)

---------

Co-authored-by: Phil Varner <phil.varner@orbitalsidekick.com>