Rework dependencies and how pytest is run (#126)
## What I'm changing

Previously, all of the dependencies across every workspace were
installed and then the tests run. This is bad because packages can be
used in code and pass tests even though the dependency is not correcly
defined in the workspace's pyproject.toml.

## How I did it

- refactored the workspace pyproject.toml's to have dev dependencies
necessary to run pytest
- created a test running script to install only the dependencies for
each workspace and run their tests.

## Checklist

- [X] Tests pass: `uv run pytest`
- [X] Checks pass: `uv run pre-commit run --all-files`
- [X] CHANGELOG is updated (if necessary)

---------

Co-authored-by: Phil Varner <phil.varner@orbitalsidekick.com>